Add means to configure if a quote should activate multi line mode or not
This change adds a flag to ReadLineFlags that, when present, will disable
the activation of multi line mode by open quotes when the "accept-line" action
is triggered (on Enter). Depending on the value of the flag, either double
or single quotes will be disregarded when parsing the input to detect open quotes.